Fix checks for Solaris x86 to work on 64 bit systems (where the os arch is amd64 instead of i386)
Get CPU cores from VMware utilities if possible

Fix race conditions in _getdmidata and _getsmbiosdata. Both subroutines
are gathering a variety of counter/collector-type variables (number of
CPU sockets, number of CPUs, physical memory, etc.) They were
incrementing the real variables for those values directly, which created
several race conditions. First, when those values are first calculated
there is a period of time where the real variables contain a partial
count, and another call to one of the subroutines to fetch one of those
values during that time would have returned that partial and incorrect
value. Secondly, if _getdmidata or _getsmbiosdata was called a second
time then all of those variables would get doubled, because we'd
re-increment the real variable again. Normally the subroutine is only
called once and all of the values are cached. But if we're unable to
collect one of the values (host_manufacturer for example, because the
manufacturer didn't populate that data on the motherboard) then
_getdmidata/_getsmbiosdata will get called each time the parent
subroutine
(get_host_manufacturer) gets called.

Implemented the --getfieldnames operation.

Cleaned up the _xml_to_hash subroutine (renaming it to _xml_to_perl in
the process, since it doesn't always return a hash now). It now uses the
data structure type hints that Rails includes in the XML it generates.