Previous commit message is in error, it refers to a different file.  It should have read:
Add support for gathering and registering system timezone

Added support for reporting the relationship between virtualization
servers and their clients. Support is specific to VMware ESX and ESXi at
the moment.

Implemented the --getfieldnames operation.

Cleaned up the _xml_to_hash subroutine (renaming it to _xml_to_perl in
the process, since it doesn't always return a hash now). It now uses the
data structure type hints that Rails includes in the XML it generates.

Fix the name under which child groups are returned as part of a node
group.